% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{popcircle}
\alias{popcircle}
\title{Ronde}
\usage{
popcircle(x, var)
}
\arguments{
\item{x}{an sf object, a simple feature collection. If x is used then spdf, df, spdfid and dfid are not.}

\item{var}{a var}
}
\description{
Plot a choropleth layer.
}
\details{
The optimum number of class depends on the number of geographical objects.
}
\examples{
library(sf)
library(popcircle)
mtq <- st_read(system.file("gpkg/mtq.gpkg", package="popcircle"))
vv <-popcircle(mtq, "POP")
par(mar = c(0,0,0,0))
plot(st_geometry(vv$circles), col = "darkseagreen", border = "darkseagreen1")
plot(st_geometry(vv$pop), col = "red", add = TRUE, lwd = 1, border = "red4")
}
