% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{popcircle}
\alias{popcircle}
\title{Proportional Circles and Shapes}
\usage{
popcircle(x, var)
}
\arguments{
\item{x}{an sf POLYGON or MULTIPOLYGON object}

\item{var}{name of the numeric field in x to get proportionalities.}
}
\value{
A list of length 2 of circles and shapes sf objects.
}
\description{
Get proportional circles and with polygons indide using a compact layout.
}
\details{
Polygons are not proportional, they are rescaled to fit the circles.
}
\examples{
library(sf)
library(popcircle)
mtq <- st_read(system.file("gpkg/mtq.gpkg", package="popcircle"))
res <- popcircle(x = mtq, var = "POP")
circles <- res$circles
shapes <- res$shapes
par(mar = c(0,0,0,0))
plot(st_geometry(circles), col = "darkseagreen", border = "darkseagreen1")
plot(st_geometry(shapes), col = "red", add = TRUE, lwd = 1, border = "red4")
if(require(cartography)){
  labelLayer(x = circles[1:10,], txt = "LIBGEO", halo = TRUE, overlap = FALSE)
}
}
